requirements.txt
celery==5.2.7
redis==4.5.4
django-celery-beat==2.5.0
Django==4.2
django-extensions==3.2.1
django-environ==0.10.0
SQLAlchemy==2.0.9
mysqlclient==2.1.1
gunicorn==20.1.0
pandas
djangorestframework
